+++ creator-node/src/components/healthCheck/healthCheckAPI.js
@@ -1,19 +1,19 @@
 const express = require('express')
 const { handleResponse, successResponse } = require('../../apiHelpers')
 const { healthCheck } = require('./healthCheck')
+const { serviceRegistry } = require('../../serviceRegistry')
 
 const router = express.Router()
 
 // Controllers
 
-const healthCheckController = async (req) => {
-  const libs = req.app.get('audiusLibs')
-  const response = healthCheck({ libs })
+const healthCheckController = async () => {
+  const response = healthCheck(serviceRegistry)
   return successResponse(response)
 }
 
 // Routes
 
-router.get('/health_check/test', handleResponse(healthCheckController))
+router.get('/health_check', handleResponse(healthCheckController))
 
 module.exports = router
